+++ src/ProjectOrigin.Register.LineProcessor/Interfaces/ICommandStepProcessor.cs
@@ -2,7 +2,7 @@ using ProjectOrigin.Register.LineProcessor.Models;
 
 namespace ProjectOrigin.Register.LineProcessor.Interfaces;
 
-public interface ICommandStepOrchestrator
+public interface ICommandStepProcessor
 {
     Task<CommandStepResult> Process(CommandStep request);
 }
