+++ src/ProjectOrigin.Register.LineProcessor/Models/CommandStepProcessorOptions.cs
@@ -1,0 +1,3 @@
+namespace ProjectOrigin.Register.LineProcessor.Models;
+
+public record CommandStepProcessorOptions(string RegistryName);
