+++ src/ProjectOrigin.Register.LineProcessor/Models/ModelLoaderOptions.cs
@@ -1,5 +1,5 @@
 using ProjectOrigin.VerifiableEventStore.Services.EventStore;
 
-namespace ProjectOrigin.Register.LineProcessor.Services;
+namespace ProjectOrigin.Register.LineProcessor.Models;
 
 public record ModelLoaderOptions(Dictionary<string, IEventStore> EventStores);
