--- src/ProjectOrigin.Register.LineProcessor/Services/RequestProcessorOptions.cs
@@ -1,3 +1,0 @@
-namespace ProjectOrigin.Register.LineProcessor.Services;
-
-public record RequestProcessorOptions(string RegistryName);
